Incident-report/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asyncio==3.4.3
backports.tarfile==1.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
docutils==0.21.2
dotenv==0.9.9
Gentoro==0.1.8
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
id==1.5.0
idna==3.10
importlib_metadata==8.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
keyring==25.6.0
langchain==0.3.21
langchain-core==0.3.45
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
langgraph==0.3.16
langgraph-checkpoint==2.0.21
langgraph-prebuilt==0.1.3
langgraph-sdk==0.1.57
langsmith==0.3.16
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.6.0
msgpack==1.1.0
nh3==0.2.20
openai==1.66.5
orjson==3.10.15
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
readme_renderer==44.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
sniffio==1.3.1
SQLAlchemy==2.0.39
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
twine==6.1.0
typing_extensions==4.12.2
urllib3==2.3.0
zipp==3.21.0
zstandard==0.23.0
